pyproject.toml
[tool.poetry]
name = "rag-assig"
version = "0.1.0"
description = ""
authors = ["devmianharoon <devmianharoon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-core = "^0.2.35"
langchain-community = "^0.2.12"
python-dotenv = "^1.0.1"
langchain-google-genai = "^1.0.10"
langchain-chroma = "^0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
